+++ vscode/test/e2e/command-edit.test.ts
@@ -34,9 +34,9 @@ test('code lenses for edit (fixup) task', async ({ page, sidebar }) => {
     const inputBox = page.getByPlaceholder(/^Enter edit instructions \(type @ to include code/)
     const instruction = 'replace hello with goodbye'
     const inputTitle = /^Edit index.html:(\d+).* with Cody$/
-    const showDiffLens = page.getByRole('button', { name: 'A Show Diff' })
+    const showDiffLens = page.getByRole('button', { name: 'Show Diff' })
     const acceptLens = page.getByRole('button', { name: 'Accept' })
-    const retryLens = page.getByRole('button', { name: 'Retry' })
+    const retryLens = page.getByRole('button', { name: 'Edit & Retry' })
     const undoLens = page.getByRole('button', { name: 'Undo' })
 
     // Wait for the input box to appear with the document name in title